refactor: moved Text API methods to extension-methods. Moved TextGizmo type to a separate file. Removed commented unused code.
docs: Added XML-documentation to DebugXTextSettings.
