--- Runtime/Gizmos/DebugX.text.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c465dc7f5c78c9549a063ecfc5e8326d
