+++ Runtime/Gizmos/Text/TextGizmo.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: af700d933ae340608724469c70e8d0fc
+timeCreated: 1740750866
